refactor(markov): move CLI parsing and help to lib (parseAndGetCommand)
